Move transform to public headers again - Replace all old get set etc. functions with get_transform for actor - Add actor flipping to transform and implement
